+++ lo2-to-sam-tester-public/.gitignore
@@ -134,5 +134,4 @@ dmypy.json
 /*.sam/
 *.class
 .nosync
-.DS_stroe
 .DS_Store
